requirements.txt
asgiref==3.3.4
Django==3.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-utils-six==2.0
django-widget-tweaks==1.4.8
Pillow==8.2.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
